refactor: 조회수 레디스 적재 기능 클래스 분리
